+++ src/main/java/io/minio/reactive/ListObjectsV2Args.java
@@ -1,0 +1,49 @@
+package io.minio.reactive;
+
+/** `ListObjectsV2Args`：列表类请求参数对象。 */
+public final class ListObjectsV2Args extends BucketArgs {
+  private final String prefix;
+  private final boolean recursive;
+
+  private ListObjectsV2Args(Builder builder) {
+    super(builder);
+    this.prefix = builder.prefix;
+    this.recursive = builder.recursive;
+  }
+
+  public String prefix() {
+    return prefix;
+  }
+
+  public boolean recursive() {
+    return recursive;
+  }
+
+  public static Builder builder() {
+    return new Builder();
+  }
+
+  public static final class Builder extends BucketArgs.AbstractBuilder<Builder> {
+    private String prefix;
+    private boolean recursive = true;
+
+    public Builder prefix(String prefix) {
+      this.prefix = prefix;
+      return this;
+    }
+
+    public Builder recursive(boolean recursive) {
+      this.recursive = recursive;
+      return this;
+    }
+
+    @Override
+    protected Builder self() {
+      return this;
+    }
+
+    public ListObjectsV2Args build() {
+      return new ListObjectsV2Args(this);
+    }
+  }
+}