+++ src/main/java/io/minio/reactive/ObjectReadArgs.java
@@ -1,0 +1,8 @@
+package io.minio.reactive;
+
+/** 对象读取参数基类。 */
+public abstract class ObjectReadArgs extends ObjectVersionArgs {
+  protected ObjectReadArgs(ObjectVersionArgs.AbstractBuilder<?> builder) {
+    super(builder);
+  }
+}