+++ src/main/java/io/minio/reactive/PutObjectAPIArgs.java
@@ -1,0 +1,8 @@
+package io.minio.reactive;
+
+/** PutObject API 参数对象基类。 */
+public abstract class PutObjectAPIArgs extends PutObjectAPIBaseArgs {
+  protected PutObjectAPIArgs(ObjectArgs.AbstractBuilder<?> builder) {
+    super(builder);
+  }
+}